Extract berkeley defs from expr and move into berkeley_columns
